Moving request handler
